DR 129362: Prep work for decision date fixes
