+++ src/applications/appeals/995/config/form.js
@@ -2,6 +2,7 @@ import { VA_FORM_IDS } from 'platform/forms/constants';
 import preSubmitInfo from 'platform/forms/preSubmitInfo';
 import FormFooter from 'platform/forms/components/FormFooter';
 import { externalServices as services } from 'platform/monitoring/DowntimeNotification';
+
 // Components
 import AddContestableIssue from '../components/AddContestableIssue';
 import ConfirmationPage from '../components/ConfirmationPage';
@@ -17,12 +18,14 @@ import SubTaskContainer from '../subtask/SubTaskContainer';
 import Summary from '../components/evidence/Summary';
 import VaDetailsEntry from '../components/evidence/VaDetailsEntry';
 import VaPrompt from '../components/evidence/VaPrompt';
+
 // Content
 import reviewErrors from '../content/reviewErrors';
 import { saveInProgress, savedFormMessages } from '../content/formMessages';
 import { title995, getSubTitle } from '../content/title';
 import submissionError from '../../shared/content/submissionError';
 import GetFormHelp from '../../shared/content/GetFormHelp';
+
 // Pages
 import addIssue from '../pages/addIssue';
 import contactInfo from '../pages/contactInformation';
@@ -49,6 +52,7 @@ import uploadPrompt from '../pages/evidence/uploadPrompt';
 import vaDetails from '../pages/evidence/vaDetails';
 import vaPrompt from '../pages/evidence/vaPrompt';
 import veteranInfo from '../pages/veteranInfo';
+
 // Utils
 import migrations from '../migrations';
 import {
