+++ src/applications/appeals/shared/tests/utils/replace.unit.spec.jsx
@@ -1,10 +1,8 @@
 import { expect } from 'chai';
-
 import {
   replaceWhitespace,
   replaceDescriptionContent,
   replaceSubmittedData,
-  fixDateFormat,
 } from '../../utils/replace';
 
 describe('replaceWhitespace', () => {
@@ -75,38 +73,5 @@ describe('replaceSubmittedData', () => {
   it('should strip out multiple whitespace & trim', () => {
     expect(replaceSubmittedData('   ')).to.eq('');
     expect(replaceSubmittedData(' a  b  c  \t  d')).to.eq('a b c d');
-  });
-});
-
-describe('fixDateFormat', () => {
-  it('should return an empty strings for empty or non-string values', () => {
-    expect(fixDateFormat()).to.eq('');
-    expect(fixDateFormat('')).to.eq('');
-    expect(fixDateFormat({})).to.eq('');
-    expect(fixDateFormat(null)).to.eq('');
-    expect(fixDateFormat(10)).to.eq('');
-  });
-  it('should return invalid dates strings', () => {
-    expect(fixDateFormat('-')).to.eq('-00-00');
-    expect(fixDateFormat('200')).to.eq('200-00-00');
-  });
-  it('should return already properly formatted date string', () => {
-    expect(fixDateFormat('2020-01-02')).to.eq('2020-01-02');
-    expect(fixDateFormat('2023-12-31')).to.eq('2023-12-31');
-    expect(fixDateFormat('2000-06-30')).to.eq('2000-06-30');
-  });
-  it('should return properly formatted date string when passed dates with no leading zero', () => {
-    expect(fixDateFormat('2020-1-2')).to.eq('2020-01-02');
-    expect(fixDateFormat('2023-10-1')).to.eq('2023-10-01');
-    expect(fixDateFormat('2000-6-30')).to.eq('2000-06-30');
-  });
-  it('should return properly formatted date string when passed dates with weird spacing', () => {
-    expect(fixDateFormat('2020--')).to.eq('2020-00-00');
-    expect(fixDateFormat('2020-1-')).to.eq('2020-01-00');
-    expect(fixDateFormat('2020- 1 - 2 ')).to.eq('2020-01-02');
-    expect(fixDateFormat('2023 - 10 - 1')).to.eq('2023-10-01');
-    expect(fixDateFormat('2000-6 - 30')).to.eq('2000-06-30');
-    expect(fixDateFormat('2000  -  6 - 30')).to.eq('2000-06-30');
-    expect(fixDateFormat('2000 \t - \t 6 \t - \t 30')).to.eq('2000-06-30');
   });
 });
