+++ src/applications/appeals/shared/utils/dates.js
@@ -1,10 +1,37 @@
-import { parse, parseISO, add, format, isValid, isToday } from 'date-fns';
-
+import { parseISODate } from 'platform/forms-system/src/js/helpers';
+import { parse, parseISO, add, format, isToday, isValid } from 'date-fns';
+import { formatInTimeZone } from 'date-fns-tz';
+import { addLeadingZero, coerceStringValue } from '.';
 import {
   FORMAT_YMD_DATE_FNS,
   FORMAT_READABLE_DATE_FNS,
+  REGEXP,
   VA_LONG_FORM_MONTHS,
 } from '../constants';
+
+const USER_TIMEZONE = Intl.DateTimeFormat().resolvedOptions().timeZone;
+
+/**
+ * Get the current timezone abbreviation (e.g., "PST", "EST", "JST")
+ * Uses date-fns-tz to properly handle DST (EDT vs EST, PDT vs PST, etc.)
+ * @returns {string} Timezone abbreviation
+ */
+export const getCurrentTimeZoneAbbr = () => {
+  const timezone = USER_TIMEZONE;
+  const now = new Date();
+
+  return formatInTimeZone(now, timezone, 'zzz');
+};
+
+/**
+ * Helper: Format date part in readable format
+ * @param {Date} date - Date to format
+ * @param {string} timezone - Target timezone (defaults to user's current timezone)
+ * @returns {string} Formatted date (e.g., "January 15, 2024")
+ */
+export const formatDatePart = (date, timezone = USER_TIMEZONE) => {
+  return formatInTimeZone(date, timezone, 'MMMM d, yyyy');
+};
 
 /**
  * parseDateToDateObj from ISO8601 or JS number date (not unix time)
@@ -153,4 +180,70 @@ export const formatDateToReadableString = date => {
   const year = format(date, 'yyyy');
 
   return `${abbreviatedMonth} ${day}, ${year}`;
+};
+
+/**
+ * Helper: Format time part in readable format
+ * @param {Date} date - Date to format
+ * @param {string} timezone - Target timezone (defaults to user's current timezone)
+ * @returns {string} Formatted time (e.g., "3:45 p.m.")
+ */
+export const formatTimePart = (date, timezone = USER_TIMEZONE) => {
+  // Use date-fns-tz to format time with proper AM/PM formatting
+  const timeString = formatInTimeZone(date, timezone, 'h:mm a');
+  // Convert "AM" to "a.m." and "PM" to "p.m." to match VA style guide
+  return timeString.replace(/AM/g, 'a.m.').replace(/PM/g, 'p.m.');
+};
+
+/**
+ * Helper: Format date with midnight in local timezone
+ * E.g., in California (UTC-8), if decision was today, UTC midnight tomorrow converts to 5:00 p.m. today
+ * Business decision: Don't allow same-day appeals, so we show stable "12:00 a.m. tomorrow" format
+ * @param {Date} date - Date to format
+ * @param {string} timezoneAbbr - Timezone abbreviation (optional, will be determined automatically)
+ * @returns {string} Formatted date with midnight time
+ */
+export const formatDateWithMidnight = (date, timezoneAbbr) => {
+  const userTimezone = USER_TIMEZONE;
+  const abbr = timezoneAbbr || getCurrentTimeZoneAbbr();
+  return `${formatDatePart(date, userTimezone)}, 12:00 a.m. ${abbr}`;
+};
+
+/**
+ * Helper: Format date with specific time in local timezone
+ * Used for showing UTC conversion times, e.g., in Japan (UTC+9), UTC midnight becomes 9:00 a.m. JST
+ * @param {Date} date - Date to format
+ * @param {string} timezoneAbbr - Timezone abbreviation (optional, will be determined automatically)
+ * @returns {string} Formatted date with specific time
+ */
+export const formatDateWithTime = (date, timezoneAbbr) => {
+  const userTimezone = USER_TIMEZONE;
+  const abbr = timezoneAbbr || getCurrentTimeZoneAbbr();
+  const datePart = formatDatePart(date, userTimezone);
+  const timePart = formatTimePart(date, userTimezone);
+  return `${datePart}, ${timePart} ${abbr}`;
+};
+
+/**
+ * Change a date string with no leading zeros (e.g. 2020-1-2) into a date with
+ * leading zeros (e.g. 2020-01-02) as expected in the schema
+ * @param {String} dateString YYYY-M-D or YYYY-MM-DD date string
+ * @returns {String} YYYY-MM-DD date string
+ */
+export const fixDateFormat = (date, yearOnly = false) => {
+  const dateString = coerceStringValue(date).replace(REGEXP.WHITESPACE, '');
+
+  if (dateString.length === 10 || dateString === '') {
+    return dateString;
+  }
+
+  // Stopgap solution to properly format a year only when provided for a VA treatment date "before 2005"
+  // Coming into this function it will have a format of {year}-01. Here we'll add a day of 01
+  // Remove this when the new date components are implemented and the fields are required
+  if (yearOnly) {
+    return `${dateString}-01`;
+  }
+
+  const { day, month, year } = parseISODate(dateString);
+  return `${year}-${addLeadingZero(month)}-${addLeadingZero(day)}`;
 };
